+++ CHANGES.md
@@ -1,6 +1,12 @@
+### 2.0.0 ###
+* Added: examples + docs
+* Added: new colorized cli logging backend
+* Added: support for named backends
+* Changed: make use of ES6 features like rest operator (requires node >= 6.4)
+* Changed: default backend is not set by default. in case no backend is set the default `cli-logger` will be used
+
 ### 1.1.0 ###
 * Added: support for multiple backends (set of upstream handlers)
-
 
 ### 1.0.0 ###
 Initial public release